Open-Source & Build Cleanup
The readme file was changed to make the aim of the project, and how to use it, much clearer. A GPL licence was also applied to the project.

A change was made to make an initial prompt regarding the project directory clearer. Numerous changes were made to project references to allow NuGet to resolve dependencies correctly.